довёл до ума работу раздела со здоровьем и добавил докстринги и комменты осталось доделать раздел с заметками

+++ main.py
@@ -5,12 +5,17 @@ from datetime import datetime
 from telebot import types
 
 bot = telebot.TeleBot(token)
-database = sqlite3.connect("database.db", check_same_thread=False)
-database_cursor = database.cursor()
+database = sqlite3.connect("database.db", check_same_thread=False)  # Подключение базы данных.
+database_cursor = database.cursor()  # Обьект курсора для управления БД.
 
 
-# new database and table creation function
 def create_table(name, request):
+    """
+    Функция добавления таблицы в БД.
+
+    :param name: название таблицы
+    :param request: SQL запрос с параметрами для новой таблицы
+    """
     with database:
         data = database.execute(f"SELECT count(*) FROM sqlite_master WHERE TYPE='table' and NAME='{name}'")
         for row in data:
@@ -19,7 +24,13 @@ def create_table(name, request):
 
 
 def change_state(user_id, state):
-    database_cursor.execute(f"SELECT * FROM people WHERE user_id={user_id}")
+    """
+    Функция изменения состояния пользователя в ходе работы бота.
+
+    :param user_id: ID пользователя
+    :param state: состояние пользователя
+    """
+    database_cursor.execute(f"SELECT * FROM people WHERE user_id={user_id}")  # Поиск информации о пользователе в БД.
     if database_cursor.fetchone() is None:
         database_cursor.execute(f"INSERT INTO people (user_id, state) VALUES ({user_id}, '{state}')")
     else:
@@ -41,21 +52,36 @@ create_table("notes",
 
 @bot.message_handler(commands=['start'])
 def start(message):
-    change_state(message.from_user.id, "start")
-    keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)  # open keyboard
+    """
+    Функция начала работы с ботом.
+
+    :param message: сообщение от пользователя
+    """
+    change_state(message.from_user.id, "start")  # Изменение состояния пользователя в боте.
+    keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)  # Открытие клавиатуры.
     keyboard.row('Состояния', 'Заметки')
     bot.send_message(message.from_user.id, f"Привет, {message.from_user.first_name}! Выбери нужный раздел:",
                      reply_markup=keyboard)
 
 
 def menu(message):
+    """
+    Функция отображения меню разделов.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "menu")
-    keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)  # open keyboard
+    keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Состояния', 'Заметки')
     bot.send_message(message.from_user.id, "Выбери нужный раздел:", reply_markup=keyboard)
 
 
 def work_with_health(message):
+    """
+    Функция отображения меню работы со здоровьем.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "work_with_health")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Показать', 'Изменить', 'Удалить')
@@ -64,6 +90,11 @@ def work_with_health(message):
 
 
 def change_health(message):
+    """
+    Функция, принимающяя информацию о здоровье от пользователя.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "change_health")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Болею', 'Не болею')
@@ -73,21 +104,31 @@ def change_health(message):
 
 
 def set_health(message):
+    """
+    Функция записи изменения здоровья в БД.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "set_health")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
-    keyboard.row('Изменить')
+    keyboard.row('Изменить', 'Удалить')
     keyboard.row('Назад', 'Меню')
-    if message.text != 'Назад':
+    if message.text in ('Болею', 'Не болею'):
         database_cursor.execute(f"UPDATE people SET health='{message.text}' WHERE user_id={message.from_user.id}")
         database.commit()
         bot.send_message(message.from_user.id, f"*Отлично! Теперь твоё состояние:* {message.text}",
                          reply_markup=keyboard, parse_mode="Markdown")
         bot.register_next_step_handler(message, action)
     else:
-        work_with_health(message)
+        action(message)
 
 
 def delete_health(message):
+    """
+    Функция удаления информации о здоровье из БД.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "delete_health")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Изменить')
@@ -101,27 +142,43 @@ def delete_health(message):
 
 
 def show_health(message):
+    """
+    Функция отображения здоровья.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "show_health")
-    keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
-    keyboard.row('Изменить')
-    keyboard.row('Назад', 'Меню')
+    keyboard_1 = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard_2 = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard_1.row('Изменить', 'Удалить')
+    keyboard_1.row('Назад', 'Меню')
+    keyboard_2.row('Изменить')
+    keyboard_2.row('Назад', 'Меню')
     if check_health(message):
         bot.send_message(message.from_user.id, f"*Твоё состояние:* {''.join(check_health(message))}",
-                         reply_markup=keyboard, parse_mode="Markdown")
+                         reply_markup=keyboard_1, parse_mode="Markdown")
     else:
-        bot.send_message(message.from_user.id, "Ты ещё не указал информацию о здоровье!", reply_markup=keyboard)
+        bot.send_message(message.from_user.id, "Ты ещё не указал информацию о здоровье!", reply_markup=keyboard_2)
 
 
 def check_health(message):
+    """
+    Функция проверки наличия информации о здоровье в БД.
+
+    :param message: сообщение от пользователя
+    :return информация о здоровье или None
+    """
     database_cursor.execute(f"SELECT health FROM people WHERE user_id={message.from_user.id}")
     data = database_cursor.fetchone()[0]
-    if data:
-        return data
-    else:
-        return False
+    return data
 
 
 def work_with_notes(message):
+    """
+    Функция отображения меню работы с заметками.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "work_with_notes")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Показать все', 'Добавить', 'Изменить')
@@ -131,6 +188,11 @@ def work_with_notes(message):
 
 
 def change_note(message):
+    """
+    Функция, принимающая ID изменяемой заметки.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "change_note")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Назад', 'Меню')
@@ -143,6 +205,11 @@ def change_note(message):
 
 
 def change_note_after_show(message):
+    """
+    Функция, принимающая ID изменяемой заметки после отображения всех заметок.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "change_note_after_show")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Назад', 'Меню')
@@ -151,25 +218,39 @@ def change_note_after_show(message):
 
 
 def change_note_text(message):
+    """
+    Функция, принимающая новый текст изменяемой заметки.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "change_note_text")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Назад', 'Меню')
-    if message.text.isdigit():
-        note_id = message.text
+    if message.text.isdigit():  # Проверка что пользователь ввёл ID заметки, а не ерунду.
+        note_id = message.text  # Сохранение введённого ID для передачи в другую функцию.
         if check_note_id(message):
             bot.send_message(message.from_user.id, "Введи новый текст заметки:", reply_markup=keyboard)
             bot.register_next_step_handler(message, change_note_in_db, note_id)
         else:
-            bot.send_message(message.from_user.id, "Введён неверный ID!", reply_markup=keyboard)
+            bot.send_message(message.from_user.id, "Введён несуществующий ID! Попробуй снова:", reply_markup=keyboard)
+            bot.register_next_step_handler(message, change_note_text)
     else:
-        action(message)
+        action(message, "Введи ID, а не текст!")
+        bot.register_next_step_handler(message, change_note_text)
 
 
 def change_note_in_db(message, note_id):
+    """
+    Функция изменения текста заметки в БД.
+
+    :param message: сообщение от пользователя
+    :param note_id: ID выбранной заметки
+    """
     change_state(message.from_user.id, "change_note_in_db")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard.row('Добавить заметку')
     keyboard.row('Назад', 'Меню')
-    if message.text != 'Назад':
+    if message.text != 'Назад' and message.text != 'Меню':
         database_cursor.execute(f"UPDATE notes SET note='{message.text}', "
                                 f"datetime='{datetime.now().strftime('%d.%m.%Y %H:%M:%S')}' WHERE id={note_id}")
         database.commit()
@@ -179,8 +260,14 @@ def change_note_in_db(message, note_id):
 
 
 def delete_note_id(message):
+    """
+    Функция, принимающая ID заметки для удаления.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "delete_note_id")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard.row('Добавить заметку')
     keyboard.row('Назад', 'Меню')
     if check_notes_user_id(message):
         print_all_notes(message)
@@ -191,20 +278,36 @@ def delete_note_id(message):
 
 
 def delete_note_id_from_db(message):
+    """
+    Функция удаления заметки из БД.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "delete_note_id_from_db")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Назад', 'Меню')
-    if check_note_id(message):
-        database_cursor.execute(f"DELETE FROM notes WHERE id={message.text}")
-        database.commit()
-        bot.send_message(message.from_user.id, "Заметка удалена!", reply_markup=keyboard)
+    if message.text.isdigit():
+        if check_note_id(message):
+            database_cursor.execute(f"DELETE FROM notes WHERE id={message.text}")
+            database.commit()
+            bot.send_message(message.from_user.id, "Заметка удалена!", reply_markup=keyboard)
+        else:
+            bot.send_message(message.from_user.id, "Введён несуществующий ID! Попробуй снова:", reply_markup=keyboard)
+            bot.register_next_step_handler(message, delete_note_id_from_db)
     else:
-        bot.send_message(message.from_user.id, "Введён неверный ID!", reply_markup=keyboard)
+        action(message, "Введи ID, а не текст!")
+        bot.register_next_step_handler(message, delete_note_id_from_db)
 
 
 def delete_all_notes(message):
+    """
+    Функция удаления всех заметок из БД.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "delete_all_notes")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
+    keyboard.row('Добавить заметку')
     keyboard.row('Назад', 'Меню')
     if check_notes_user_id(message):
         database_cursor.execute(f"DELETE FROM notes WHERE user_id={message.from_user.id}")
@@ -215,6 +318,11 @@ def delete_all_notes(message):
 
 
 def show_notes(message):
+    """
+    Функция отображения всех заметок.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "show_notes")
     keyboard_1 = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard_1.row('Изменить заметку')
@@ -231,24 +339,35 @@ def show_notes(message):
 
 
 def check_notes_user_id(message):
+    """
+    Функция проверки наличия заметок у пользователя в БД.
+
+    :param message: сообщение от пользователя
+    :return найденные заметки или None
+    """
     database_cursor.execute(f"SELECT * FROM notes WHERE user_id={message.from_user.id}")
     data = database_cursor.fetchone()
-    if data:
-        return True
-    else:
-        return False
+    return data
 
 
 def check_note_id(message):
+    """
+    Функция проверки наличия заметки с указанным ID в БД.
+
+    :param message: сообщение от пользователя
+    :return найденные заметки или None
+    """
     database_cursor.execute(f"SELECT * FROM notes WHERE id={message.text}")
     data = database_cursor.fetchone()
-    if data:
-        return True
-    else:
-        return False
+    return data
 
 
 def print_all_notes(message):
+    """
+    Функция печати всех заметок пользователя.
+
+    :param message: сообщение от пользователя
+    """
     database_cursor.execute(f"SELECT id, datetime, note FROM notes WHERE user_id={message.from_user.id}")
     for data in database_cursor.fetchall():
         note_id = f"*ID заметки:* {data[0]}"
@@ -258,6 +377,11 @@ def print_all_notes(message):
 
 
 def add_note(message):
+    """
+    Функция принятия текста заметки для добавления.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "add_note")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Назад', 'Меню')
@@ -266,17 +390,26 @@ def add_note(message):
 
 
 def write_note(message):
+    """
+    Функция сохранения заметки в БД.
+
+    :param message: сообщение от пользователя
+    """
     change_state(message.from_user.id, "write_note")
     keyboard = telebot.types.ReplyKeyboardMarkup(resize_keyboard=True)
     keyboard.row('Добавить заметку')
     keyboard.row('Назад', 'Меню')
-    database_cursor.execute(f"INSERT INTO notes (user_id, datetime, note) "
-                            f"VALUES ('{message.from_user.id}', '{datetime.now().strftime('%d.%m.%Y %H:%M:%S')}', "
-                            f"'{message.text}')")
-    database.commit()
-    bot.send_message(message.from_user.id, "Заметка сохранена!", reply_markup=keyboard)
+    if message.text != 'Назад' and message.text != 'Меню':
+        database_cursor.execute(f"INSERT INTO notes (user_id, datetime, note) "
+                                f"VALUES ('{message.from_user.id}', '{datetime.now().strftime('%d.%m.%Y %H:%M:%S')}', "
+                                f"'{message.text}')")
+        database.commit()
+        bot.send_message(message.from_user.id, "Заметка сохранена!", reply_markup=keyboard)
+    else:
+        action(message)
 
 
+# Словарь, хранящий все варианты состояний пользователя и действий на кнопки в конкретных состояниях.
 actions = {
     ("start", "Меню"): start,
     ("start", "Состояния"): work_with_health,
@@ -293,20 +426,27 @@ actions = {
     ("work_with_health", "Изменить"): change_health,
     ("delete_health", "Изменить"): change_health,
     ("work_with_health", "Удалить"): delete_health,
+    ("set_health", "Удалить"): delete_health,
+    ("set_health", "Болею"): set_health,
+    ("set_health", "Не болею"): set_health,
+    ("show_health", "Удалить"): delete_health,
     ("work_with_health", "Показать"): show_health,
 
     ("change_note", "Назад"): work_with_notes,
     ("change_note_text", "Назад"): work_with_notes,
     ("delete_note_id", "Назад"): work_with_notes,
+    ("delete_note_id", "Добавить заметку"): add_note,
     ("delete_note_id_from_db", "Назад"): work_with_notes,
     ("delete_all_notes", "Назад"): work_with_notes,
+    ("delete_all_notes", "Добавить заметку"): add_note,
     ("show_notes", "Назад"): work_with_notes,
     ("show_notes", "Изменить заметку"): change_note_after_show,
     ("show_notes", "Добавить заметку"): add_note,
     ("show_notes", "Удалить все"): delete_all_notes,
     ("show_notes", "Удалить по ID"): delete_note_id,
     ("change_note_after_show", "Назад"): work_with_notes,
-    ("change_note_in_db", "Назад"): change_note_after_show,
+    ("change_note_in_db", "Назад"): work_with_notes,
+    ("change_note_in_db", "Добавить заметку"): add_note,
     ("add_note", "Назад"): work_with_notes,
     ("write_note", "Назад"): work_with_notes,
     ("write_note", "Добавить заметку"): add_note,
@@ -319,7 +459,13 @@ actions = {
 
 
 @bot.message_handler(content_types=['text'])
-def action(message):
+def action(message, text="Выбери нужный пункт!"):
+    """
+    Функция определения действий для бота в зависимости от нажатой пользователем кнопки.
+
+    :param message: сообщение от пользователя
+    :param text: текст предупреждения пользователю
+    """
     if message.text == 'Меню':
         menu(message)
     elif message.text == '/start':
@@ -327,7 +473,10 @@ def action(message):
     else:
         database_cursor.execute(f"SELECT state FROM people WHERE user_id={message.from_user.id}")
         state = ''.join(database_cursor.fetchone())
-        actions[(state, message.text)](message)
+        if (state, message.text) in actions:
+            actions[(state, message.text)](message)
+        else:
+            bot.send_message(message.from_user.id, text)
 
 
 bot.infinity_polling()
